New Editor (#3)
* Files relocation

* Editor frame inhereits from DFrame now + fixed starfall's file browser causing errors

* Switched to StarfallFileBrowser

* Got rid of old frame

* Started work on tabeditors

* Added missing directory

* Editor now includes/registers its own files

* TabEditor > TabHandler

* More work on TabHandlers

* Added example of tab interface

* Implementing interface in tab_wire and sfframe

* New method for tab interface

* Some work on ACE

* Added 1px spacing between tabs and changed interface.. back

* Some color fixes

* Added missing file + sessions are working

* Fixed ace unparenting when closing wrong tab

* Forgot about debug msg

* Fixed editor initializing multiple times

* Sessions update, they should work now

* Fixed shortcuts while using ACE

* Whitespace stuff

* Fixes,changes, control panel mainly

* Removed unused stuff

* Forgot debug.. again

* Cleaning up

* Fixed convar name + ace is now default

* Fixed right-clicking on existing chip

* Removed debug font print

* Editor reload command added

* Syntax coloring

* Wire settings

* Fixed #187
